V.2.0 backend initial set up
