+++ api/middleware/auth.js
@@ -1,6 +1,6 @@
-const { User } = require('../models/User');
+import User from '../models/User.js'
 
-let auth = (req, res, next) => {
+export const auth = (req, res, next) => {
   let token = req.cookies.w_auth;
 
   User.findByToken(token, (err, user) => {
@@ -17,4 +17,3 @@ let auth = (req, res, next) => {
   });
 };
 
-module.exports = { auth };
