+++ api/models/User.js
@@ -1,8 +1,8 @@
-const mongoose = require('mongoose');
-const bcrypt = require('bcrypt');
+import mongoose from 'mongoose';
+import bcrypt from 'bcrypt'
+import jwt from 'jsonwebtoken'
+import moment from 'moment';
 const saltRounds = 10;
-const jwt = require('jsonwebtoken');
-const moment = require("moment");
 
 const userSchema = mongoose.Schema({
     name: {
@@ -96,6 +96,4 @@ userSchema.statics.findByToken = function (token, cb) {
     })
 }
 
-const User = mongoose.model('User', userSchema);
-
-module.exports = { User }
+export default mongoose.model('User', userSchema);
