+++ client/src/components/utils/Paypal.js
@@ -28,7 +28,7 @@ export default class Paypal extends React.Component {
         // Document on Paypal's currency code: https://developer.paypal.com/docs/classic/api/currency_codes/
  
         const client = {
-            sandbox:    'YOUR-SANDBOX-APP-ID',
+            sandbox:    'AeIKnMneB18izQug6Cdm7hZVYP2dZ3W59FeuIGES60DiZxw6mg0kcm26Npxp3t351ifUTF8wRChvEGwc',
             production: 'YOUR-PRODUCTION-APP-ID',
         }
         // In order to get production's app-ID, you will have to send your app to Paypal for approval first
