--- server/config/prod.js
@@ -1,3 +1,0 @@
-module.exports = {
-    mongoURI:process.env.MONGO_URI
-}
