--- server/index.js
@@ -1,52 +1,0 @@
-const express = require("express");
-const app = express();
-const path = require("path");
-const cors = require('cors')
-
-const bodyParser = require("body-parser");
-const cookieParser = require("cookie-parser");
-
-const config = require("./config/key");
-
-const mongoose = require("mongoose");
-const connect = mongoose.connect(config.mongoURI)
-  .then(() => console.log('MongoDB Connected...'))
-  .catch(err => console.log(err));
-
-app.use(cors())
-
-//to not get any deprecation warning or error
-//support parsing of application/x-www-form-urlencoded post data
-app.use(bodyParser.urlencoded({ extended: true }));
-//to get json data
-// support parsing of application/json type post data
-app.use(bodyParser.json());
-app.use(cookieParser());
-
-app.use('/api/users', require('./routes/users'));
-app.use('/api/product', require('./routes/product'));
-
-//use this to show the image you have in node js server to client (react js)
-//https://stackoverflow.com/questions/48914987/send-image-path-from-node-js-express-server-to-react-client
-app.use('/uploads', express.static('uploads'));
-
-// Serve static assets if in production
-if (process.env.NODE_ENV === "production") {
-
-  // Set static folder   
-  // All the javascript and css files will be read and served from this folder
-  app.use(express.static("client/build"));
-
-  // index.html for all page routes    html or routing and naviagtion
-  app.get("*", (req, res) => {
-    res.sendFile(path.resolve(__dirname, "../client", "build", "index.html"));
-  });
-}
-app.get('/', function (req, res) {
-  res.send('Hello World')
-})
-const port = process.env.PORT || 5000
-
-app.listen(port, () => {
-  console.log(`Server Listening on ${port}`)
-});
